+++ src/MiNET/MiNET/Entities/Passive/Wolf.cs
@@ -66,7 +66,7 @@ namespace MiNET.Entities.Passive
 			Behaviors.Add(new RandomLookaroundBehavior(this));
 		}
 
-		public override void DoInteraction(byte actionId, Player player)
+		public override void DoInteraction(int actionId, Player player)
 		{
 			if (IsTamed)
 			{
